Split runner artifact read tests
