+++ crates/webcodex-runner/src/main_tests.rs
@@ -102,6 +102,8 @@ fn quic_client_config() -> QuicClientConfig {
 
 #[path = "main_tests/agent_config.rs"]
 mod agent_config;
+#[path = "main_tests/artifact_read.rs"]
+mod artifact_read;
 #[path = "main_tests/config_reload.rs"]
 mod config_reload;
 #[path = "main_tests/file_read.rs"]
@@ -1195,18 +1197,6 @@ fn structured_delete_project_files_errors_do_not_leak_absolute_paths() {
     assert!(!error.contains(&tmp.path().to_string_lossy().to_string()));
 }
 
-fn fake_zip_eocd_with_entries(entries: u16) -> Vec<u8> {
-    let mut bytes = b"PK\x05\x06".to_vec();
-    bytes.extend_from_slice(&[0, 0]); // disk number
-    bytes.extend_from_slice(&[0, 0]); // central directory disk
-    bytes.extend_from_slice(&entries.to_le_bytes());
-    bytes.extend_from_slice(&entries.to_le_bytes());
-    bytes.extend_from_slice(&[0, 0, 0, 0]); // central directory size
-    bytes.extend_from_slice(&[0, 0, 0, 0]); // central directory offset
-    bytes.extend_from_slice(&[0, 0]); // comment length
-    bytes
-}
-
 fn append_fake_zip_entry(
     bytes: &mut Vec<u8>,
     central_directory: &mut Vec<u8>,
@@ -1307,31 +1297,6 @@ fn fake_ooxml_zip(
     bytes
 }
 
-fn fake_zip_central_entry_offset(bytes: &[u8], expected_name: &str) -> usize {
-    let eocd = bytes
-        .windows(4)
-        .rposition(|window| window == b"PK\x05\x06")
-        .unwrap();
-    let entry_count = u16::from_le_bytes(bytes[eocd + 10..eocd + 12].try_into().unwrap());
-    let mut cursor = u32::from_le_bytes(bytes[eocd + 16..eocd + 20].try_into().unwrap()) as usize;
-    for _ in 0..entry_count {
-        assert_eq!(&bytes[cursor..cursor + 4], b"PK\x01\x02");
-        let name_len =
-            u16::from_le_bytes(bytes[cursor + 28..cursor + 30].try_into().unwrap()) as usize;
-        let extra_len =
-            u16::from_le_bytes(bytes[cursor + 30..cursor + 32].try_into().unwrap()) as usize;
-        let comment_len =
-            u16::from_le_bytes(bytes[cursor + 32..cursor + 34].try_into().unwrap()) as usize;
-        let name_start = cursor + 46;
-        let name_end = name_start + name_len;
-        if &bytes[name_start..name_end] == expected_name.as_bytes() {
-            return cursor;
-        }
-        cursor = name_end + extra_len + comment_len;
-    }
-    panic!("missing fake ZIP central entry {expected_name}");
-}
-
 fn artifact_upload_temp_paths(
     root: &Path,
     artifact_path: &str,
@@ -1391,61 +1356,6 @@ fn assert_no_upload_temp_files(root: &Path, artifact_path: &str) {
         "legacy .pd upload temp files remained in {}",
         parent.display()
     );
-}
-
-#[test]
-fn file_save_project_artifact_writes_binary_and_blocks_overwrite() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-    let path = "artifacts/imports/tiny.png";
-    let content_base64 = base64::Engine::encode(
-        &base64::engine::general_purpose::STANDARD,
-        [0x89, b'P', b'N', b'G'],
-    );
-    let payload = serde_json::json!({
-        "path": path,
-        "content_base64": content_base64,
-        "mime_type": "image/png",
-        "overwrite": false,
-        "max_bytes": 1024,
-    });
-
-    let out = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_save_project_artifact",
-            path,
-            payload.clone(),
-        ),
-    ));
-
-    assert_eq!(out["path"], path);
-    assert_eq!(out["bytes_written"], 4);
-    assert_eq!(out["mime_type"], "image/png");
-    assert_eq!(out["sha256"].as_str().unwrap().len(), 64);
-    assert_eq!(
-        std::fs::read(tmp.path().join(path)).unwrap(),
-        vec![0x89, b'P', b'N', b'G']
-    );
-    let parent = tmp.path().join("artifacts/imports");
-    assert!(
-        !directory_contains_name_prefix(&parent, ".wc-artifact-"),
-        "atomic artifact temp file should not remain"
-    );
-    assert!(
-        !directory_contains_name_prefix(&parent, ".pd-artifact-"),
-        "legacy .pd artifact temp file should not remain"
-    );
-
-    let out2 = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(tmp.path(), "file_save_project_artifact", path, payload),
-    ));
-    assert!(out2["error"]
-        .as_str()
-        .unwrap()
-        .contains("overwrite is false"));
 }
 
 #[test]
@@ -1510,477 +1420,6 @@ fn file_save_and_upload_begin_accept_office_mimes() {
         let upload_id = upload["upload_id"].as_str().unwrap();
         assert_upload_temp_files_exist(tmp.path(), &upload_path, upload_id);
     }
-}
-
-#[test]
-fn file_read_project_artifact_metadata_counts_zip_without_extracting() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-    let zip_path = tmp.path().join("sample.zip");
-    std::fs::write(&zip_path, fake_zip_eocd_with_entries(2)).unwrap();
-
-    let out = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            "sample.zip",
-            serde_json::json!({"path": "sample.zip", "max_bytes": 1024}),
-        ),
-    ));
-
-    assert_eq!(out["mime_type"], "application/zip");
-    assert_eq!(out["archive_entries_count"], 2);
-    assert!(
-        out["modified_at"].as_u64().unwrap() > 0,
-        "modified_at should be unix timestamp seconds"
-    );
-    assert!(!tmp.path().join("a.txt").exists());
-    assert!(!tmp.path().join("b.txt").exists());
-}
-
-#[test]
-fn file_read_project_artifact_detects_ooxml_mime_from_package_content() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-    let cases = [
-        (
-            "sample.docx",
-            "word/document.xml",
-            "application/vnd.openxmlformats-officedocument.wordprocessingml.document.main+xml",
-            "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-        ),
-        (
-            "sample.pptx",
-            "ppt/presentation.xml",
-            "application/vnd.openxmlformats-officedocument.presentationml.presentation.main+xml",
-            "application/vnd.openxmlformats-officedocument.presentationml.presentation",
-        ),
-        (
-            "sample.xlsx",
-            "xl/workbook.xml",
-            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet.main+xml",
-            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-        ),
-    ];
-
-    for (path, main_part, main_content_type, expected_mime) in cases {
-        let bytes = fake_ooxml_zip(main_part, main_content_type, false);
-        std::fs::write(tmp.path().join(path), &bytes).unwrap();
-        let metadata = line_edit_json(handle_file_request(
-            &policy,
-            &json_file_op_request(
-                tmp.path(),
-                "file_read_project_artifact_metadata",
-                path,
-                serde_json::json!({"path": path, "max_bytes": 64 * 1024}),
-            ),
-        ));
-        assert_eq!(metadata["mime_type"], expected_mime, "{path}");
-        assert!(metadata.get("archive_entries_count").is_none(), "{path}");
-
-        let read = line_edit_json(handle_file_request(
-            &policy,
-            &json_file_op_request(
-                tmp.path(),
-                "file_read_project_artifact",
-                path,
-                serde_json::json!({
-                    "path": path,
-                    "offset": 0,
-                    "length": 16,
-                    "max_file_bytes": 64 * 1024,
-                }),
-            ),
-        ));
-        assert_eq!(read["mime_type"], expected_mime, "{path}");
-        assert_eq!(read["file_bytes"], bytes.len(), "{path}");
-        assert_eq!(read["bytes_returned"], 16, "{path}");
-    }
-}
-
-#[test]
-fn file_read_project_artifact_does_not_trust_ooxml_extension_or_malformed_package() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-
-    std::fs::write(tmp.path().join("spoof.docx"), fake_zip_eocd_with_entries(0)).unwrap();
-    let spoof = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            "spoof.docx",
-            serde_json::json!({"path": "spoof.docx", "max_bytes": 1024}),
-        ),
-    ));
-    assert_eq!(spoof["mime_type"], "application/zip");
-
-    std::fs::write(tmp.path().join("not-a-zip.docx"), b"plain bytes").unwrap();
-    let non_zip = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            "not-a-zip.docx",
-            serde_json::json!({"path": "not-a-zip.docx", "max_bytes": 1024}),
-        ),
-    ));
-    assert!(non_zip["mime_type"].is_null());
-
-    let malformed = fake_ooxml_zip(
-        "word/document.xml",
-        "application/vnd.openxmlformats-officedocument.wordprocessingml.document.main+xml",
-        true,
-    );
-    std::fs::write(tmp.path().join("broken.docx"), malformed).unwrap();
-    let broken = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            "broken.docx",
-            serde_json::json!({"path": "broken.docx", "max_bytes": 64 * 1024}),
-        ),
-    ));
-    assert_eq!(broken["mime_type"], "application/zip");
-}
-
-#[test]
-fn file_read_project_artifact_rejects_ooxml_main_part_with_invalid_local_structure() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-    let main_part = "word/document.xml";
-    let main_content_type =
-        "application/vnd.openxmlformats-officedocument.wordprocessingml.document.main+xml";
-
-    let mut invalid_offset = fake_ooxml_zip(main_part, main_content_type, false);
-    let central_entry = fake_zip_central_entry_offset(&invalid_offset, main_part);
-    invalid_offset[central_entry + 42..central_entry + 46].copy_from_slice(&1_u32.to_le_bytes());
-    std::fs::write(tmp.path().join("bad-offset.docx"), invalid_offset).unwrap();
-    let bad_offset = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            "bad-offset.docx",
-            serde_json::json!({"path": "bad-offset.docx", "max_bytes": 64 * 1024}),
-        ),
-    ));
-    assert_eq!(bad_offset["mime_type"], "application/zip");
-
-    let mut mismatched_name = fake_ooxml_zip(main_part, main_content_type, false);
-    let central_entry = fake_zip_central_entry_offset(&mismatched_name, main_part);
-    let local_offset = u32::from_le_bytes(
-        mismatched_name[central_entry + 42..central_entry + 46]
-            .try_into()
-            .unwrap(),
-    ) as usize;
-    let local_name_start = local_offset + 30;
-    assert_eq!(
-        &mismatched_name[local_name_start..local_name_start + main_part.len()],
-        main_part.as_bytes()
-    );
-    mismatched_name[local_name_start] = b'v';
-    std::fs::write(tmp.path().join("bad-local-name.docx"), mismatched_name).unwrap();
-    let bad_name = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            "bad-local-name.docx",
-            serde_json::json!({"path": "bad-local-name.docx", "max_bytes": 64 * 1024}),
-        ),
-    ));
-    assert_eq!(bad_name["mime_type"], "application/zip");
-}
-
-#[test]
-fn file_read_project_artifact_reads_binary_chunks() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-    let bytes = [0, 159, 146, 150, b'a', b'b', b'c', b'd'];
-    std::fs::write(tmp.path().join("data.bin"), bytes).unwrap();
-
-    let first = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact",
-            "data.bin",
-            serde_json::json!({"path": "data.bin", "offset": 0, "length": 4, "max_file_bytes": 1024}),
-        ),
-    ));
-    assert_eq!(first["file_bytes"], bytes.len());
-    assert!(first["sha256"]
-        .as_str()
-        .is_some_and(|value| value.len() == 64));
-    assert!(first.get("mime_type").is_some());
-    assert_eq!(first["offset"], 0);
-    assert_eq!(first["bytes_returned"], 4);
-    assert_eq!(first["next_offset"], 4);
-    assert_eq!(first["truncated"], true);
-    assert_eq!(first["eof"], false);
-    assert_eq!(
-        first["content_base64"],
-        base64::Engine::encode(&base64::engine::general_purpose::STANDARD, &bytes[..4])
-    );
-
-    let second = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact",
-            "data.bin",
-            serde_json::json!({"path": "data.bin", "offset": 4, "length": 20, "max_file_bytes": 1024}),
-        ),
-    ));
-    assert_eq!(second["sha256"], first["sha256"]);
-    assert_eq!(second["offset"], 4);
-    assert_eq!(second["bytes_returned"], bytes.len() - 4);
-    assert_eq!(second["next_offset"], bytes.len());
-    assert_eq!(second["truncated"], false);
-    assert_eq!(second["eof"], true);
-    assert_eq!(
-        second["content_base64"],
-        base64::Engine::encode(&base64::engine::general_purpose::STANDARD, &bytes[4..])
-    );
-
-    let at_eof = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact",
-            "data.bin",
-            serde_json::json!({"path": "data.bin", "offset": bytes.len(), "length": 4, "max_file_bytes": 1024}),
-        ),
-    ));
-    assert_eq!(at_eof["bytes_returned"], 0);
-    assert_eq!(at_eof["next_offset"], bytes.len());
-    assert_eq!(at_eof["truncated"], false);
-    assert_eq!(at_eof["eof"], true);
-}
-
-#[test]
-fn file_read_project_artifact_export_chunk_reads_only_requested_segments() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-    let bytes = vec![0x5a; 70 * 1024];
-    std::fs::write(tmp.path().join("export.bin"), &bytes).unwrap();
-
-    let first = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_export_chunk",
-            "export.bin",
-            serde_json::json!({
-                "path": "export.bin",
-                "expected_file_bytes": bytes.len(),
-                "offset": 0,
-                "length": 64 * 1024
-            }),
-        ),
-    ));
-    assert_eq!(first["file_bytes"], bytes.len());
-    assert_eq!(first["offset"], 0);
-    assert_eq!(first["bytes_returned"], 64 * 1024);
-    assert_eq!(first["next_offset"], 64 * 1024);
-    assert_eq!(first["truncated"], true);
-    assert_eq!(first["eof"], false);
-    assert!(first.get("sha256").is_none());
-    assert!(first.get("mime_type").is_none());
-    let first_bytes = base64::Engine::decode(
-        &base64::engine::general_purpose::STANDARD,
-        first["content_base64"].as_str().unwrap(),
-    )
-    .unwrap();
-    assert_eq!(first_bytes, bytes[..64 * 1024]);
-
-    let final_chunk = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_export_chunk",
-            "export.bin",
-            serde_json::json!({
-                "path": "export.bin",
-                "expected_file_bytes": bytes.len(),
-                "offset": 64 * 1024,
-                "length": 64 * 1024
-            }),
-        ),
-    ));
-    assert_eq!(final_chunk["offset"], 64 * 1024);
-    assert_eq!(final_chunk["bytes_returned"], 6 * 1024);
-    assert_eq!(final_chunk["next_offset"], bytes.len());
-    assert_eq!(final_chunk["truncated"], false);
-    assert_eq!(final_chunk["eof"], true);
-    let final_bytes = base64::Engine::decode(
-        &base64::engine::general_purpose::STANDARD,
-        final_chunk["content_base64"].as_str().unwrap(),
-    )
-    .unwrap();
-    assert_eq!(final_bytes, bytes[64 * 1024..]);
-
-    let wrong_size = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_export_chunk",
-            "export.bin",
-            serde_json::json!({
-                "path": "export.bin",
-                "expected_file_bytes": bytes.len() - 1,
-                "offset": 0,
-                "length": 1
-            }),
-        ),
-    ));
-    assert_eq!(wrong_size["error_kind"], "snapshot_changed");
-
-    let ten_mib = 10 * 1024 * 1024;
-    let boundary = std::fs::File::create(tmp.path().join("boundary.bin")).unwrap();
-    boundary.set_len(ten_mib as u64).unwrap();
-    let boundary_read = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_export_chunk",
-            "boundary.bin",
-            serde_json::json!({
-                "path": "boundary.bin",
-                "expected_file_bytes": ten_mib,
-                "offset": ten_mib - 1,
-                "length": 64 * 1024
-            }),
-        ),
-    ));
-    assert_eq!(boundary_read["bytes_returned"], 1);
-    assert_eq!(boundary_read["next_offset"], ten_mib);
-    assert_eq!(boundary_read["eof"], true);
-
-    let above_whole_payload =
-        std::fs::File::create(tmp.path().join("above-whole-payload.bin")).unwrap();
-    above_whole_payload.set_len((ten_mib + 1) as u64).unwrap();
-    let above_whole_payload_read = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_export_chunk",
-            "above-whole-payload.bin",
-            serde_json::json!({
-                "path": "above-whole-payload.bin",
-                "expected_file_bytes": ten_mib + 1,
-                "offset": ten_mib,
-                "length": 1
-            }),
-        ),
-    ));
-    assert_eq!(above_whole_payload_read["bytes_returned"], 1);
-    assert_eq!(above_whole_payload_read["next_offset"], ten_mib + 1);
-    assert_eq!(above_whole_payload_read["eof"], true);
-
-    let export_max = 256 * 1024 * 1024;
-    let max_file = std::fs::File::create(tmp.path().join("export-max.bin")).unwrap();
-    max_file.set_len(export_max as u64).unwrap();
-    let max_read = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_export_chunk",
-            "export-max.bin",
-            serde_json::json!({
-                "path": "export-max.bin",
-                "expected_file_bytes": export_max,
-                "offset": export_max - 1,
-                "length": 1
-            }),
-        ),
-    ));
-    assert_eq!(max_read["bytes_returned"], 1);
-    assert_eq!(max_read["next_offset"], export_max);
-    assert_eq!(max_read["eof"], true);
-
-    let too_large = std::fs::File::create(tmp.path().join("export-too-large.bin")).unwrap();
-    too_large.set_len((export_max + 1) as u64).unwrap();
-    let rejected = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_export_chunk",
-            "export-too-large.bin",
-            serde_json::json!({
-                "path": "export-too-large.bin",
-                "expected_file_bytes": export_max + 1,
-                "offset": 0,
-                "length": 1
-            }),
-        ),
-    ));
-    assert!(rejected["error"].as_str().unwrap().contains("maximum"));
-}
-
-#[test]
-fn file_read_project_artifact_metadata_streams_above_whole_payload_limit() {
-    let tmp = tempfile::tempdir().unwrap();
-    let policy = project_policy(tmp.path());
-    let ten_mib = 10 * 1024 * 1024;
-    let export_max = 256 * 1024 * 1024;
-    let path = "large-export.pdf";
-    let file = std::fs::File::create(tmp.path().join(path)).unwrap();
-    file.set_len((ten_mib + 1) as u64).unwrap();
-
-    let metadata = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            path,
-            serde_json::json!({
-                "path": path,
-                "max_bytes": export_max,
-                "allow_missing": false
-            }),
-        ),
-    ));
-    assert!(metadata.get("error").is_none(), "metadata: {metadata:?}");
-    assert_eq!(metadata["bytes"], ten_mib + 1);
-    assert_eq!(metadata["mime_type"], "application/pdf");
-    assert_eq!(metadata["sha256"].as_str().unwrap().len(), 64);
-
-    let whole_payload_bound = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            path,
-            serde_json::json!({
-                "path": path,
-                "max_bytes": ten_mib,
-                "allow_missing": false
-            }),
-        ),
-    ));
-    assert!(whole_payload_bound["error"]
-        .as_str()
-        .unwrap()
-        .contains("too large"));
-
-    let invalid_max = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            tmp.path(),
-            "file_read_project_artifact_metadata",
-            path,
-            serde_json::json!({
-                "path": path,
-                "max_bytes": export_max + 1,
-                "allow_missing": false
-            }),
-        ),
-    ));
-    assert!(invalid_max["error"].as_str().unwrap().contains("maximum"));
 }
 
 #[test]
@@ -2827,80 +2266,6 @@ fn file_artifact_upload_abort_rejects_wrong_ids_and_cleans_only_temp() {
 }
 
 #[cfg(unix)]
-#[test]
-fn file_project_artifact_ops_reject_symlink_escape() {
-    let root = tempfile::tempdir().unwrap();
-    let outside_dir = tempfile::tempdir().unwrap();
-    let outside = outside_dir.path().join("outside.bin");
-    std::fs::write(&outside, b"outside-secret-content").unwrap();
-    std::os::unix::fs::symlink(&outside, root.path().join("leak.bin")).unwrap();
-    let mut policy = project_policy(root.path());
-    policy.allowed_roots.push(outside_dir.path().to_path_buf());
-
-    let read = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            root.path(),
-            "file_read_project_artifact",
-            "leak.bin",
-            serde_json::json!({"path":"leak.bin","offset":0,"length":8,"max_file_bytes":1024}),
-        ),
-    ));
-    assert_eq!(read["error"], "artifact path escapes project root");
-    assert!(!read.to_string().contains("outside-secret-content"));
-
-    let export_chunk = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            root.path(),
-            "file_read_project_artifact_export_chunk",
-            "leak.bin",
-            serde_json::json!({
-                "path":"leak.bin",
-                "expected_file_bytes":8,
-                "offset":0,
-                "length":8
-            }),
-        ),
-    ));
-    assert_eq!(export_chunk["error"], "artifact path escapes project root");
-    assert!(!export_chunk.to_string().contains("outside-secret-content"));
-
-    let metadata = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            root.path(),
-            "file_read_project_artifact_metadata",
-            "leak.bin",
-            serde_json::json!({"path":"leak.bin","max_bytes":1024}),
-        ),
-    ));
-    assert_eq!(metadata["error"], "artifact path escapes project root");
-    assert!(!metadata.to_string().contains("outside-secret-content"));
-
-    let save = line_edit_json(handle_file_request(
-        &policy,
-        &json_file_op_request(
-            root.path(),
-            "file_save_project_artifact",
-            "leak.bin",
-            serde_json::json!({
-                "path":"leak.bin",
-                "content_base64":"bmV3",
-                "mime_type":"text/plain",
-                "overwrite":true,
-                "max_bytes":1024
-            }),
-        ),
-    ));
-    assert_eq!(save["error"], "refusing to overwrite symlink artifact path");
-    assert_eq!(
-        std::fs::read(&outside).expect("outside file remains readable"),
-        b"outside-secret-content"
-    );
-    assert!(!save.to_string().contains("outside-secret-content"));
-}
-
 #[test]
 fn file_write_project_file_creates_parent_dirs_and_reports_hash() {
     let tmp = tempfile::tempdir().unwrap();
